health_care_advisor/requirements.txt
flask==2.3.3
flask-cors==5.0.0
python-dotenv==1.0.0
requests==2.28.1